renamed paths and files
