+++ home_Agent.html
@@ -168,8 +168,10 @@
         <div class="col-md-6">
           <div class="card p-5" style="background-color: #087830">
             <div class="card-body">
-              <h2 style="color: white;" class="mb-4">Input Client Info</h2>			  
-              <form action = "http://localhost:8080/upperlimit/clientadded.php" method = "post">	  
+              <h2 style="color: white;" class="mb-4">Input Client Info</h2>			
+
+
+              <form action = "home_Agent_clientadded.php" method = "post">	  
                 <div class="form-group"> <label>First Name</label>
                   <input type="text" name="firstname" class="form-control" placeholder="Enter client's first name" required> </div>
 				<div class="form-group"> <label>Last Name</label>
